+++ examples/wcp-retry-idempotency/main.go
@@ -18,7 +18,7 @@ import (
 	"os"
 	"strings"
 
-	axonflow "github.com/getaxonflow/axonflow-sdk-go/v8"
+	axonflow "github.com/getaxonflow/axonflow-sdk-go/v9"
 )
 
 func main() {